Move content loader to shared ui
