+++ idea/frontend/src/features/voucher/ui/voucher-table/voucher-table.tsx
@@ -2,7 +2,7 @@ import { HexString } from '@gear-js/api';
 import { useAccount } from '@gear-js/react-hooks';
 
 import { ReactComponent as VoucherPlaceholderSVG } from 'features/voucher/assets/voucher-placeholder.svg';
-import { ContentLoader } from 'pages/program/ui/contentLoader';
+import { ContentLoader } from 'shared/ui/contentLoader';
 import { BulbBlock, BulbStatus } from 'shared/ui/bulbBlock';
 import { Table, TableRow } from 'shared/ui/table';
