+++ idea/frontend/src/shared/ui/contentLoader/ContentLoader.module.scss
