+++ idea/frontend/src/shared/ui/contentLoader/ContentLoader.tsx
