+++ idea/frontend/src/shared/ui/contentLoader/index.ts
